Refactor measurements command
